.Last.lib {R.utils}	R Documentation

Undo changed done by this package when detached

Description

Undo changed done by this package when detached. Reverts .Last() to the function that existed before this package was attached.

Usage


.Last.lib(libpath)

Arguments

libpath
a character string giving the complete path to the package.

Value

Returns nothing.

Author(s)

Henrik Bengtsson

[Package R.utils version 2.12.3 Index]