Abstract class SmartComments

Package: R.utils
Class SmartComments

Object
~~|
~~+--SmartComments

Directly known subclasses:
LComments, VComments

public abstract static class SmartComments
extends Object

Abstract class SmartComments.

SmartComments(letter=NA, ...)

Arguments

letter

A single character.

...

Not used.

Fields and Methods

Methods:

compile	Preprocess a vector of code lines.

convertComment	Converts a single smart comment to R code.

parse	Parses one single smart comment.

reset	Resets a SmartComments compiler.

validate	Validates the compiled lines.

Methods inherited from Object:
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, names, objectSize, print, save

Details

A "smart" source-code comment is an R comment, which start with a '#', but is followed by a single letter, then a single symbol and a second '#' and then an option character string, and there must not be any code before the comment on the same line. In summary, a smart comment line has format: <white spaces>#<letter><symbol># <some text>.

Example code with two smart comments (VComments):

   x <- 2
   #V1# threshold=-1
   #Vc# A v-comment log message
   cat("Hello world")
 

which after compilation becomes

   x <- 2
   verbose <- Verbose(threshold=-1)
   if (verbose) { cat(verbose, "A v-comment log message"); }
   cat("Hello world")
 

Author

Henrik Bengtsson

See also