A status bar at the R prompt that can be updated

Package: R.utils
Class TextStatusBar

Object
~~|
~~+--TextStatusBar

Directly known subclasses:

public static class TextStatusBar
extends Object

A status bar at the R prompt that can be updated.

TextStatusBar(fmt=paste("%-", getOption("width") - 1, "s", sep = ""), ...)

Arguments

fmt

A character format string to be used by sprintf(). Default is a left-aligned string of full width.

...

Named arguments to be passed to sprintf() together with the format string.

Fields and Methods

Methods:

flush	Flushes the output.

getLabel	Gets the current value of a label.

newline	Writes a newline.

popMessage	Adds a message above the status bar.

setLabel	Sets the value of a label.

setLabels	Sets new values of given labels.

update	Updates the status bar (visually).

updateLabels	Sets the new values of given labels and updates the status bar.

Methods inherited from Object:
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, names, objectSize, print, save

Details

A label with name hfill can be used for automatic horizontal filling. It must be numeric and be immediate before a string label such that a hfill label and the following string label together specifies an sprintf format such as "%*-s". The value of hfill will be set such that the resulting status bar has width equal to getOption("width")-1 (the reason for the -1 is to prevent the text status bar from writing into the next line). If more than one hfill label is used their widths will be uniformly distributed. Left over spaces will be distributed between hfill labels with initial values of one.

Author

Henrik Bengtsson

Examples

# - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
# Read all HTML files in the base package
# - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
path <- system.file(package="base")
files <- list.files(path, recursive=TRUE, full.names=TRUE)
files <- files[sapply(files, FUN=isFile)]
nfiles <- length(files)

cat(sprintf("Reading %d files in %s:\n", nfiles, path))
#> Reading 24 files in /home/hb/shared/software/CBI/R-4.2.1-gcc9/lib/R/library/base:

# Create a status bar with four labels
sb <- TextStatusBar("File: %-*s [%3.0f%% %7.0f bytes %-8s]",
                hfill=1, file="", progress=0, nbytes=0L, time="")

nbytes <- 0L
for (kk in seq_len(nfiles)) {
  file <- files[kk]

  # Update the status bar
  if (sb) {
    setLabel(sb, "progress", 100*kk/nfiles)
    if (kk %% 10 == 1 || kk == nfiles)
      setLabel(sb, "file", substr(basename(file), 1, 44))

    size <- file.info(file)$size
    # popMessage() calls update() too
    popMessage(sb, sprintf("Processing %s (%.2fkB)",
                                       basename(file), size/1024))
    flush(sb)
  }

  # Read the file
  bfr <- readBin(file, what="raw", n=size)
  nbytes <- nbytes + size

  # Emulate a slow process
  if (interactive()) Sys.sleep(rexp(1, rate=60))

  # Update the status bar
  if (sb) {
    setLabel(sb, "nbytes", nbytes)
    setLabel(sb, "time", format(Sys.time(), "%H:%M:%S"))
    update(sb)
  }
}
#> Processing CITATION (0.93kB)
#> File: CITATION                                    [  4%       0 bytes         ]File: CITATION                                    [  4%     956 bytes 18:25:34]                                                                               Processing DESCRIPTION (0.35kB)
#> File: CITATION                                    [  8%     956 bytes 18:25:34]File: CITATION                                    [  8%    1310 bytes 18:25:34]                                                                               Processing INDEX (23.81kB)
#> File: CITATION                                    [ 12%    1310 bytes 18:25:34]File: CITATION                                    [ 12%   25689 bytes 18:25:34]                                                                               Processing Rd.rds (21.70kB)
#> File: CITATION                                    [ 17%   25689 bytes 18:25:34]File: CITATION                                    [ 17%   47905 bytes 18:25:34]                                                                               Processing demo.rds (0.26kB)
#> File: CITATION                                    [ 21%   47905 bytes 18:25:34]File: CITATION                                    [ 21%   48172 bytes 18:25:34]                                                                               Processing features.rds (0.12kB)
#> File: CITATION                                    [ 25%   48172 bytes 18:25:34]File: CITATION                                    [ 25%   48294 bytes 18:25:34]                                                                               Processing hsearch.rds (22.39kB)
#> File: CITATION                                    [ 29%   48294 bytes 18:25:34]File: CITATION                                    [ 29%   71217 bytes 18:25:34]                                                                               Processing links.rds (13.03kB)
#> File: CITATION                                    [ 33%   71217 bytes 18:25:34]File: CITATION                                    [ 33%   84559 bytes 18:25:34]                                                                               Processing package.rds (0.54kB)
#> File: CITATION                                    [ 38%   84559 bytes 18:25:34]File: CITATION                                    [ 38%   85115 bytes 18:25:34]                                                                               Processing Rprofile (7.11kB)
#> File: CITATION                                    [ 42%   85115 bytes 18:25:34]File: CITATION                                    [ 42%   92396 bytes 18:25:34]                                                                               Processing base (5.03kB)
#> File: base                                        [ 46%   92396 bytes 18:25:34]File: base                                        [ 46%   97548 bytes 18:25:34]                                                                               Processing base.rdb (1049.37kB)
#> File: base                                        [ 50%   97548 bytes 18:25:34]File: base                                        [ 50% 1172106 bytes 18:25:34]                                                                               Processing base.rdx (13.46kB)
#> File: base                                        [ 54% 1172106 bytes 18:25:34]File: base                                        [ 54% 1185891 bytes 18:25:34]                                                                               Processing error.catching.R (1.15kB)
#> File: base                                        [ 58% 1185891 bytes 18:25:34]File: base                                        [ 58% 1187070 bytes 18:25:34]                                                                               Processing is.things.R (4.69kB)
#> File: base                                        [ 62% 1187070 bytes 18:25:34]File: base                                        [ 62% 1191868 bytes 18:25:34]                                                                               Processing recursion.R (2.07kB)
#> File: base                                        [ 67% 1191868 bytes 18:25:34]File: base                                        [ 67% 1193983 bytes 18:25:34]                                                                               Processing scoping.R (1.52kB)
#> File: base                                        [ 71% 1193983 bytes 18:25:34]File: base                                        [ 71% 1195543 bytes 18:25:34]                                                                               Processing AnIndex (34.19kB)
#> File: base                                        [ 75% 1195543 bytes 18:25:34]File: base                                        [ 75% 1230558 bytes 18:25:34]                                                                               Processing aliases.rds (12.21kB)
#> File: base                                        [ 79% 1230558 bytes 18:25:34]File: base                                        [ 79% 1243065 bytes 18:25:34]                                                                               Processing base.rdb (2223.93kB)
#> File: base                                        [ 83% 1243065 bytes 18:25:34]File: base                                        [ 83% 3520368 bytes 18:25:34]                                                                               Processing base.rdx (8.91kB)
#> File: base.rdx                                    [ 88% 3520368 bytes 18:25:34]File: base.rdx                                    [ 88% 3529496 bytes 18:25:34]                                                                               Processing paths.rds (3.04kB)
#> File: base.rdx                                    [ 92% 3529496 bytes 18:25:34]File: base.rdx                                    [ 92% 3532614 bytes 18:25:34]                                                                               Processing 00Index.html (147.39kB)
#> File: base.rdx                                    [ 96% 3532614 bytes 18:25:34]File: base.rdx                                    [ 96% 3683537 bytes 18:25:34]                                                                               Processing R.css (1.69kB)
#> File: R.css                                       [100% 3683537 bytes 18:25:34]File: R.css                                       [100% 3685272 bytes 18:25:34]
setLabel(sb, "file", "<done>")
update(sb)
#> File: <done>                                      [100% 3685272 bytes 18:25:34]
cat("\n")
#>