addFinalizerToLast {R.utils}	R Documentation

Modifies .Last() to call 'finalizeSession()

Description

Modifies .Last() to call 'finalizeSession() before calling the default .Last() function.

Note that .Last() is not guaranteed to be called when the R session finished. For instance, the user may quit R by calling quit(runLast=FALSE) or run R in batch mode.

Note that this function is called when the R.utils package is loaded.

Usage


## Default S3 method:
addFinalizerToLast(...)

Arguments

...
Not used.

Value

Returns (invisibly) TRUE if .Last() was modified, otherwise FALSE.

Author(s)

Henrik Bengtsson

See Also

onSessionExit().

[Package R.utils version 2.12.3 Index]