callHooks.function {R.utils}	R Documentation

Call hook functions

Description

Call hook functions.

Usage


## S3 method for class 'function'
callHooks(hooks, ...)

Arguments

hooks
A function or a list of hook functions or names of such.

...
Argument passed to each hook function.

Value

Returns (invisibly) a list that is named with hook names, if possible. Each element in the list is in turn a list with three element: fcn is the hook function called, result is its return value, and exception is the exception caught or NULL.

Author(s)

Henrik Bengtsson

See Also

See callHooks() to call hook function by name.

[Package R.utils version 2.12.3 Index]