cat.Verbose {R.utils}	R Documentation

Concatenates and prints objects if above threshold

Description

Concatenates and prints objects if above threshold. The output is indented according to *enter()/*exit() calls.

Usage


## S3 method for class 'Verbose'
cat(this, ..., sep="", newline=TRUE, level=this$defaultLevel, timestamp=this$.timestamp)

Arguments

...
Objects to be passed to cat().

sep
The default separator character string.

newline
If TRUE, a newline is added at the end, otherwise not.

level
A numeric value to be compared to the threshold.

timestamp
A logical indicating if output should start with a timestamp, or not.

Value

Returns nothing.

Author(s)

Henrik Bengtsson

See Also

*timestampOn() and timestampOff(). For more information see Verbose.

[Package R.utils version 2.12.3 Index]