Assert$check {R.utils}	R Documentation

Static method asserting that a generic condition is true

Description

Static method asserting that a generic condition is true.

Usage


## Static method (use this):
## Assert$check(condition, message=NULL, ...)

## Don't use the below:
## S3 method for class 'Assert'
check(static, condition, message=NULL, ...)

Arguments

condition
A condition that should return TRUE if ok, or something else if not.

message
The error message to be reported on failure. If NULL, a message is automatically created.

...
Not used.

Value

Returns (invisibly) TRUE, or throws an exception.

Author(s)

Henrik Bengtsson

See Also

For more information see Assert.

[Package R.utils version 2.12.3 Index]