countLines {R.utils}	R Documentation

Counts the number of lines in a text file

Description

Counts the number of lines in a text file by counting the number of occurrences of platform-independent newlines (CR, LF, and CR+LF [1]), including a last line with neither. An empty file has zero lines.

Usage


## Default S3 method:
countLines(file, chunkSize=5e+07, ...)

Arguments

file
A connection or a pathname.

chunkSize
The number of bytes read in each chunk.

...
Not used.

Details

Both compressed and non-compressed files are supported.

Value

Returns an non-negative integer.

Author(s)

Henrik Bengtsson

References

[1] Page Newline, Wikipedia, July 2008. https://en.wikipedia.org/wiki/Newline

Examples


pathname <- system.file("NEWS.md", package="R.utils");
n <- countLines(pathname);
n2 <- length(readLines(pathname));
stopifnot(n == n2);

[Package R.utils version 2.12.3 Index]