detachPackage {R.utils}	R Documentation

Detaches packages by name

Description

Detaches packages by name, if loaded.

Usage


## Default S3 method:
detachPackage(pkgname, ...)

Arguments

pkgname
A character vector of package names to be detached.

...
Not used.

Value

Returns (invisibly) a named logical vector indicating whether each package was detached or not.

Author(s)

Henrik Bengtsson

See Also

detach().

[Package R.utils version 2.12.3 Index]