draw.density {R.utils}	R Documentation

Draws a density curve

Description

Draws a density curve along one of the sides of the current plotting region.

Usage


## S3 method for class 'density'
draw(object, side=1, height=0.2, offset=0, scale=c("absolute", "relative"), xtrim=NULL,
  xpd=TRUE, ...)

Arguments

side
An integer specifying which side to draw along. See mtext for details.

height
A numeric scalar specifying the "height" of the curve, where height means the maximum height of the density. that is, how much the zero-density position should be shifted.

scale
A character specifying the scale of the curve, which can be either absolute or relative.

xpd
If TRUE, the curve is not clipped, cf. par.

...
Not used.

Value

Returns the drawn 'density' object (with the 'x' and 'y' coordinates as plotted).

Author(s)

Henrik Bengtsson

See Also

See density for estimating densities. Internally *swapXY() may be used.

[Package R.utils version 2.12.3 Index]