enter.Verbose {R.utils}	R Documentation

Writes a message and indents the following output

Description

Writes a message and indents the following output. The output is indented according to *enter()/*exit() calls.

Usage


  ## S3 method for class 'Verbose'
enter(this, ..., indent=this$indentStep, sep="", suffix="...", level=this$defaultLevel)
  ## S3 method for class 'Verbose'
enterf(this, fmtstr, ..., indent=this$indentStep, sep="", suffix="...",
  level=this$defaultLevel)

Arguments

fmtstr
An sprintf() format string, which together with ... constructs the message.

...
Objects to be passed to *cat() (or sprintf()).

indent
The number of characters to add to the indentation.

sep
The default separator character string.

suffix
A character string to be appended to the end of the message.

level
A numeric value to be compared to the threshold.

Value

Returns nothing.

Author(s)

Henrik Bengtsson

See Also

For more information see Verbose.

[Package R.utils version 2.12.3 Index]