exit.Verbose {R.utils}	R Documentation

Writes a message and unindents the following output

Description

Writes a message and unindents the following output. The output is indented according to *enter()/*exit() calls.

Usage


## S3 method for class 'Verbose'
exit(this, ..., indent=-this$indentStep, sep="", suffix="...done", level=NULL)

Arguments

...
Objects to be passed to *cat(). If not specified the message used in the corresponding *enter() call is used.

indent
The number of characters to be removed from the indentation.

sep
The default separator character string.

suffix
A character string to be appended to the end of the message.

level
A numeric value to be compared to the threshold.

Value

Returns nothing.

Author(s)

Henrik Bengtsson

See Also

For more information see Verbose.

[Package R.utils version 2.12.3 Index]