System$findGhostscript {R.utils}	R Documentation

Searches for a Ghostview executable on the current system

Description

Searches for a Ghostview executable on the current system.

Usage


## Static method (use this):
## System$findGhostscript(updateRGSCMD=TRUE, firstOnly=TRUE, force=FALSE, ...)

## Don't use the below:
## S3 method for class 'System'
findGhostscript(static, updateRGSCMD=TRUE, firstOnly=TRUE, force=FALSE, ...)

Arguments

updateRGSCMD
If TRUE and Ghostscript is found, then the system environment variable R_GSCMD is set to the (first) path found.

firstOnly
If TRUE, only the first executable is returned.

force
If TRUE, existing R_GSCMD is ignored, otherwise not.

...
Not used.

Value

Returns a character vector of full and normalized pathnames where Ghostscript executables are found.

Author(s)

Henrik Bengtsson

References

[1] How to use Ghostscript, Ghostscript, 2022 https://ghostscript.com/docs/9.55.0/Use.htm
[2] Environment variable, Wikipedia, 2013. https://en.wikipedia.org/wiki/Environment_variable
[3] Environment.SpecialFolder Enumeration, Microsoft, 2013. https://docs.microsoft.com/en-us/dotnet/api/system.environment.specialfolder

See Also

For more information see System.

Examples

## Not run: 
  print(System$findGhostscript())

## End(Not run)

[Package R.utils version 2.12.3 Index]