Settings$findSettings {R.utils}	R Documentation

Searches for the settings file in one or several directories

Description

Searches for the settings file in one or several directories.

Usage


## Static method (use this):
## Settings$findSettings(basename, paths=c(".", "~"), ...)

## Don't use the below:
## S3 method for class 'Settings'
findSettings(static, basename, paths=c(".", "~"), ...)

Arguments

basename
A character string of the basename of the settings file.

paths
A vector of character string specifying the directories to be searched.

...
Not used.

Value

Returns the absolute pathname (character string) of the first settings file found, otherwise NULL.

Author(s)

Henrik Bengtsson

See Also

For more information see Settings.

[Package R.utils version 2.12.3 Index]