getAbsolutePath {R.utils}	R Documentation

Gets the absolute pathname string

Description

Gets the absolute pathname string.

Usage


## Default S3 method:
getAbsolutePath(pathname, workDirectory=getwd(), expandTilde=FALSE, ...)

Arguments

pathname
A character string of the pathname to be converted into an absolute pathname.

workDirectory
A character string of the current working directory.

expandTilde
If TRUE, tilde (~) is expanded to the corresponding directory, otherwise not.

...
Not used.

Details

This method will replace replicated slashes ('/') with a single one, except for the double forward slashes prefixing a Microsoft Windows UNC (Universal Naming Convention) pathname.

Value

Returns a character string of the absolute pathname.

Author(s)

Henrik Bengtsson

See Also

isAbsolutePath().

[Package R.utils version 2.12.3 Index]