Arguments$getFilename {R.utils}	R Documentation

Gets and validates a filename

Description

Gets and validates a filename.

Usage


## Static method (use this):
## Arguments$getFilename(filename, nchar=c(1, 128), class=c("safe"), .name=NULL,
##   .type="filename", ...)

## Don't use the below:
## S3 method for class 'Arguments'
getFilename(static, filename, nchar=c(1, 128), class=c("safe"), .name=NULL,
  .type="filename", ...)

Arguments

filename
A character string.

nchar
An integer vector of length two specifying the range of valid filename lengths.

class
A character string specifying the class of valid filenames.

.name
The name of the argument validated.

.type
Not used.

...
Not used.

Details

When argument class="safe", the following 86 ASCII characters are allowed in filenames:


     #$
    @ABCDEFGHIJKLMNOPQRSTUVWXYZ[]^_  (31)
    `abcdefghijklmnopqrstuvwxyz{|}~  (31)
  

This class of filenames has been extensively tested on for cross-platform support on Microsoft Windows, macOS, and various Unix flavors.

Value

Returns a character string if filename is valid, otherwise an exception is thrown.

Missing values

If filename is a missing value, then an exception is thrown.

Author(s)

Henrik Bengtsson

References

[1] Microsoft, Naming Files, Paths, and Namespaces, 2018. https://learn.microsoft.com/en-us/windows/win32/fileio/naming-a-file.

See Also

For more information see Arguments.

[Package R.utils version 2.12.3 Index]