Arguments$getLogicals {R.utils}	R Documentation

Coerces to a logical vector and validates

Description

Coerces to a logical vector and validates.

Usage


## Static method (use this):
## Arguments$getLogicals(x, ..., disallow=c("NA", "NaN"), coerce=FALSE, .name=NULL)

## Don't use the below:
## S3 method for class 'Arguments'
getLogicals(static, x, ..., disallow=c("NA", "NaN"), coerce=FALSE, .name=NULL)

Arguments

x
A vector.

disallow
A character vector specifying disallowed value sets after coercing, i.e. "NA".

...
Arguments passed to @method "getVector".

.name
A character string for name used in error messages.

Value

Returns a numeric vector.

Author(s)

Henrik Bengtsson

See Also

For more information see Arguments.

[Package R.utils version 2.12.3 Index]