getOption.Options {R.utils}	R Documentation

Gets an option

Description

Gets an option in the options tree structure or return a default value.

Usage


## S3 method for class 'Options'
getOption(this, pathname=NULL, defaultValue=NULL, ...)

Arguments

pathname
A single or a vector of character strings specifying the paths to the options to be queried. By default the complete options structure is returned.

defaultValue
The default value to be returned, if option is missing. If multiple options are queried at the same times, multiple default values may be specified as a vector or a list.

...
Not used.

Value

If a single option is queried, a single value is returned. If a vector of options are queried, a list of values are returned. For non-existing options, the default value is returned.

Author(s)

Henrik Bengtsson

See Also

*hasOption(). *setOption(). For more information see Options.

[Package R.utils version 2.12.3 Index]