Arguments$getReadablePathname {R.utils}	R Documentation

Gets a readable pathname

Description

Gets a readable pathname.

Usage


## Static method (use this):
## Arguments$getReadablePathname(file=NULL, path=NULL, mustExist=TRUE, absolute=FALSE,
##   adjust=c("none", "url"), ...)

## Don't use the below:
## S3 method for class 'Arguments'
getReadablePathname(static, file=NULL, path=NULL, mustExist=TRUE, absolute=FALSE,
  adjust=c("none", "url"), ...)

Arguments

file
A character string specifying the file.

path
A character string specifying the path.

mustExist
If TRUE, the pathname must exists and be readable, otherwise an exception is thrown. If FALSE, no such test is performed.

absolute
If TRUE, the absolute pathname is returned.

...
Not used.

Value

Returns a character string of the absolute pathname of the file.

Missing values

If file or path is NA and mustExist is FALSE, then (character) NA is returned, otherwise an exception is thrown.

Windows

If a too long pathname is detected on Windows, an informative warning is given. The maximum number of symbols in a Windows pathname is 256, including file separators '/' or '\', but excluding the drive letter, and initial file separator (e.g. 'C:/'), and the string terminator ('\0'), cf. 'MSDN - Naming a File or Directory', Microsoft. In R, the limit is one symbol less, i.e. 255.

Author(s)

Henrik Bengtsson

See Also

*getWritablePathname() filePath. For more information see Arguments.

[Package R.utils version 2.12.3 Index]