Arguments$getVector {R.utils}	R Documentation

Validates a vector

Description

Validates a vector by checking its length (number of elements).

Usage


## Static method (use this):
## Arguments$getVector(x, length=NULL, .name=NULL, ...)

## Don't use the below:
## S3 method for class 'Arguments'
getVector(static, x, length=NULL, .name=NULL, ...)

Arguments

x
A single vector.

length
A numeric vector of length two or more. If two, it is the minimum and maximum length of x. Otherwise, it is the set of possible lengths of x.

.name
A character string for name used in error messages.

...
Not used.

Value

Returns the same vector, if it is valid. Otherwise an exception is thrown.

Author(s)

Henrik Bengtsson

See Also

For more information see Arguments.

[Package R.utils version 2.12.3 Index]