Arguments$getVerbose {R.utils}	R Documentation

Coerces to Verbose object

Description

Coerces to Verbose object.

Usage


## Static method (use this):
## Arguments$getVerbose(verbose, defaultThreshold=-1, useNullVerbose=TRUE, ...,
##   .name=NULL)

## Don't use the below:
## S3 method for class 'Arguments'
getVerbose(static, verbose, defaultThreshold=-1, useNullVerbose=TRUE, ..., .name=NULL)

Arguments

verbose
A single object. If a Verbose, it is immediately returned. If a numeric value, it is used as the threshold. Otherwise the object is coerced to a logical value and if TRUE, the threshold is defaultThreshold.

defaultThreshold
A numeric value for the default threshold, if verbose was interpreted as a logical value.

useNullVerbose
If verbose can be interpreted as FALSE, return a NullVerbose object if TRUE.

...
Passed to the constructor of Verbose.

.name
A character string for name used in error messages.

Value

Returns a Verbose (or a NullVerbose) object.

Author(s)

Henrik Bengtsson

See Also

For more information see Arguments.

[Package R.utils version 2.12.3 Index]