inAnyInterval.numeric {R.utils}	R Documentation

Checks if a set of values are inside one or more intervals

Description

Checks if a set of values are inside one or more intervals.

Usage


## S3 method for class 'numeric'
inAnyInterval(...)

Arguments

...
Arguments passed to *mapToIntervals().

Value

Returns a logical vector.

Author(s)

Henrik Bengtsson

See Also

mapToIntervals().

[Package R.utils version 2.12.3 Index]