Assert$inheritsFrom {R.utils}	R Documentation

Static method asserting that an object inherits from of a certain class

Description

Static method asserting that an object inherits from of a certain class.

Usage


## Static method (use this):
## Assert$inheritsFrom(object, class, ...)

## Don't use the below:
## S3 method for class 'Assert'
inheritsFrom(static, object, class, ...)

Arguments

object
Object to be checked.

class
Name of class.

...
Not used.

Value

Returns (invisibly) TRUE, or throws an exception.

Author(s)

Henrik Bengtsson

See Also

For more information see Assert.

[Package R.utils version 2.12.3 Index]