Install R packages by name or URL

Install R packages by name or URL.

# S3 method for default
installPackages(pkgs, types="auto", repos=getOption("repos"), ..., destPath=".",
  cleanup=TRUE)

Arguments

pkgs

A character vector specifying the names and/or the URLs of the R packages to be installed.

types

A character vector of corresponding package types.

repos

A character vector of package repository URLs.

...

Additional arguments passed to install.packages.

destPath

Path where any downloaded files are saved.

cleanup

If TRUE, downloaded and successfully installed package files are removed, otherwise not.

Value

Returns nothing.

Limitations

This method cannot install any packages that are already in use. Certain packages are always in use when calling this method, e.g. R.methodsS3, R.oo, and R.utils.

Author

Henrik Bengtsson

Examples

if (FALSE) {
 installPackages("R.rsp")
 installPackages("https://cran.r-project.org/src/contrib/Archive/R.rsp/R.rsp_0.8.2.tar.gz")
 installPackages("https://cran.r-project.org/bin/windows/contrib/4.0/R.rsp_0.44.0.zip")
}