intervalsToSeq.matrix {R.utils}	R Documentation

Generates a vector of indices from a matrix of intervals

Description

Generates a vector of indices from a matrix of intervals.

Usage


## S3 method for class 'matrix'
intervalsToSeq(fromTo, sort=FALSE, unique=FALSE, ...)

Arguments

fromTo
An Nx2 integer matrix.

sort
If TRUE, the returned indices are ordered.

unique
If TRUE, the returned indices are unique.

...
Not used.

Author(s)

Henrik Bengtsson

See Also

seqToIntervals().

Examples

## Not run: See example(seqToIntervals)

[Package R.utils version 2.12.3 Index]