isEof.connection {R.utils}	R Documentation

Checks if the current file position for a connection is at the 'End of File'

Description

Checks if the current file position for a connection is at the 'End of File'.

Usage


## S3 method for class 'connection'
isEof(con, ...)

Arguments

con
A connection.

...
Not used.

Details

Internally seek() is used, which according to to the R help is discouraged on Windows. However, after many years of large-scale testing on various Windows versions and file systems we have yet to experience issues with using seek() on Windows.

Value

Returns a logical.

Author(s)

Henrik Bengtsson

See Also

For more information see connection.

[Package R.utils version 2.12.3 Index]