isFile {R.utils}	R Documentation

Checks if the file specification is a file

Description

Checks if the file specification is a file.

Usage


## Default S3 method:
isFile(pathname, ...)

Arguments

pathname
A character string of the pathname to be checked.

...
Not used.

Value

Returns TRUE if the file specification is a file, otherwise FALSE is returned.

Symbolic links

This function follows symbolic links (also on Windows) and returns a value based on the link target (rather than the link itself).

Author(s)

Henrik Bengtsson

See Also

To check if it is a directory see isDirectory(). Internally file.info() is used. See also file_test.

[Package R.utils version 2.12.3 Index]