isOpen.character {R.utils}	R Documentation

Checks if there is an open connection to a file

Description

Checks if there is an open connection to a file.

Usage


## S3 method for class 'character'
isOpen(pathname, rw=c("read", "write"), ...)

Arguments

pathname
An character vector.

rw
A character vector. If "read", a file is considered to be open if there exist an open connection that can read from that file. If "write", a file is considered to be open if there exist an open connection that can write to that file. Both these values may be specified.

...
Not used.

Value

Returns a logical vector indicating for each file whether there exists an open file connection or not.

Author(s)

Henrik Bengtsson

See Also

See isOpen() in connections. showConnections().

[Package R.utils version 2.12.3 Index]