Assert$isScalar {R.utils}	R Documentation

Static method asserting that an object is a single value

Description

Static method asserting that an object is a single value.

Usage


## Static method (use this):
## Assert$isScalar(x, ...)

## Don't use the below:
## S3 method for class 'Assert'
isScalar(static, x, ...)

Arguments

x
Object to be checked.

...
Not used.

Value

Returns (invisibly) TRUE, or throws an exception.

Author(s)

Henrik Bengtsson

See Also

For more information see Assert.

[Package R.utils version 2.12.3 Index]