isVisible.Verbose {R.utils}	R Documentation

Checks if a certain verbose level will be shown or not

Description

Checks if a certain verbose level will be shown or not.

Usage

## S3 method for class 'Verbose'
isVisible(this, level=this$defaultLevel, ...)

Arguments

level
A numeric value to be compared to the threshold.

...
Not used.

Value

Returns TRUE, if given level is greater than (not equal to) the current threshold, otherwise FALSE is returned.

Author(s)

Henrik Bengtsson

See Also

*getThreshold() and *setThreshold(). For more information see Verbose.

[Package R.utils version 2.9.1 Index]