lastModified {R.utils}	R Documentation

Gets the time when the file was last modified

Description

Gets the time when the file was last modified. The time is returned as a POSIXct object.

Usage


## Default S3 method:
lastModified(pathname, ...)

Arguments

pathname
A character string of the pathname to be checked.

...
Not used.

Value

Returns POSIXct object specifying when the file was last modified. If the file does not exist or it is a directory, 0 is returned.

Symbolic links

This function follows symbolic links (also on Windows) and returns a value based on the link target (rather than the link itself).

Author(s)

Henrik Bengtsson

See Also

Internally file.info() is used.

[Package R.utils version 2.12.3 Index]