loadAnywhere.Settings {R.utils}	R Documentation

Loads settings from file

Description

Loads settings from file. If the settings was read from file, they are by default written back to the same file. If this was not the case, it defaults to the settings file in the home directory of the current user.

Usage


## S3 method for class 'Settings'
loadAnywhere(static, file=NULL, ..., verbose=FALSE)

Arguments

file
A character string or a connection from which settings should be read. If NULL, the settings file is searched for by *findSettings().

...
Arguments passed to *findSettings().

verbose
If TRUE, verbose information is written while reading, otherwise not.

Value

Returns a Settings object if file was successfully read, otherwise NULL.

Author(s)

Henrik Bengtsson

See Also

*saveAnywhere(). For more information see Settings.

[Package R.utils version 2.12.3 Index]