loadObject {R.utils}	R Documentation

Method to load object from a file or a connection

Description

Method to load object from a file or a connection, which previously have been saved using saveObject().

Usage


## Default S3 method:
loadObject(file, path=NULL, format=c("auto", "xdr", "rds"), ...)

Arguments

file
A filename or connection to read the object from.

path
The path where the file exists.

format
File format.

...
Not used.

Details

The main difference from this method and load() in the base package, is that this one returns the object read rather than storing it in the global environment by its default name. This makes it possible to load objects back using any variable name.

Value

Returns the saved object.

Author(s)

Henrik Bengtsson

See Also

saveObject() to save an object to file. Internally load() is used. See also loadToEnv(). See also saveRDS().

[Package R.utils version 2.12.3 Index]