moveInSearchPath {R.utils}	R Documentation

Moves a environment in the search path to another position

Description

Moves a environment in the search path to another position.

Usage


## Default S3 method:
moveInSearchPath(from, to, where=c("before", "after"), ...)

Arguments

from
An integer specifying the position of the environment to be moved, or a character specifying the name of the environment to be moved.

to
The destination position like the from argument.

where
A character string specify where in relation to the destination position the environment should be moved.

...
Not used.

Details

It is not possible to move the first environment in the search path, i.e. the so called global environment.

Value

Returns (invisibly) the name of the environment moved, if it was moved, otherwise NULL.

Author(s)

Henrik Bengtsson

See Also

search().

Examples


  # Make package 'utils' come behind 'datasets' in the search path
  moveInSearchPath("package:utils", "package:datasets", where="after")

[Package R.utils version 2.12.3 Index]