mpager {R.utils}	R Documentation

A \"pager\" function that outputs to standard error

Description

A \"pager\" function that outputs to standard error and is compatible with file.show().

Usage


mpager(files, header=NULL, title="R Information", delete.file=FALSE)

Arguments

files
A character vector of K pathnames.

header
A character vector of K headers.

title
A character string.

delete.file
If TRUE, the files are deleted after displayed, otherwise not.

Value

Returns nothing.

Author(s)

Henrik Bengtsson

See Also

file.show() and argument pager.

[Package R.utils version 2.12.3 Index]