nullfile {R.utils}	R Documentation

Gets the pathname or a connection to the NULL device on the current platform

Description

Gets the pathname or a connection to the NULL device on the current platform.

Usage


  nullfile()
  nullcon()

Value

nullfile() returns a character string, which is "/dev/null" except on Windows where it is "nul:". nullcon() returns a newly opened (binary) connection to the NULL device - make sure to close it when no longer needed.

Author(s)

Henrik Bengtsson

See Also

In R (>= 3.6.0), there exists base::nullfile(), which is identical to R.utils::nullfile().

[Package R.utils version 2.12.3 Index]