parse.SmartComments {R.utils}	R Documentation

Parses one single smart comment

Description

Parses one single smart comment.

Usage


## S3 method for class 'SmartComments'
parse(this, lines, currLine, ..., letter=this$letter, pattern=NULL)

Arguments

lines
A character vector of lines of code containing smart comments (only).

currLine
The line number on which to smart smart comment begins.

...
Not used.

letter
The letter of the smart comment. Available to avoid lookup at every line.

pattern
The pattern of the smart comment.

Value

Returns a list structure.

Author(s)

Henrik Bengtsson

See Also

For more information see SmartComments.

[Package R.utils version 2.12.3 Index]