popBackupFile {R.utils}	R Documentation

Drops a backup suffix from the backup pathname

Description

Drops a backup suffix from the backup pathname and, by default, restores an existing backup file accordingly by renaming it.

Usage


## Default S3 method:
popBackupFile(filename, path=NULL, suffix=".bak", isFile=TRUE,
  onMissing=c("ignore", "error"), drop=TRUE, ..., verbose=FALSE)

Arguments

filename
The filename of the backup file.

path
The path of the file.

suffix
The suffix of the filename to be dropped.

isFile
If TRUE, the backup file must exist and will be renamed. If FALSE, it is only the pathname string that will be modified. For details, see below.

onMissing
A character string specifying what to do if the backup file does not exist.

drop
If TRUE, the backup file will be dropped in case the original file already exists or was successfully restored.

...
Not used.

verbose
A logical or Verbose.

Value

Returns the pathname with the backup suffix dropped.

Author(s)

Henrik Bengtsson

See Also

See pushBackupFile() for more details and an example.

[Package R.utils version 2.12.3 Index]