popTemporaryFile {R.utils}	R Documentation

Drops a temporary suffix from the temporary pathname

Description

Drops a temporary suffix from the temporary pathname and, by default, renames an existing temporary file accordingly.

Usage


## Default S3 method:
popTemporaryFile(filename, path=NULL, suffix=".tmp", isFile=TRUE, ..., verbose=FALSE)

Arguments

filename
The filename of the temporary file.

path
The path of the temporary file.

suffix
The suffix of the temporary filename to be dropped.

isFile
If TRUE, the temporary file must exist and will be renamed. If FALSE, it is only the pathname string that will be modified. For details, see below.

...
Not used.

verbose
A logical or Verbose.

Details

If isFile is FALSE, the pathname where the suffix of the temporary pathname has been dropped is returned. If isFile is TRUE, the temporary file is renamed. Then, if the temporary file does not exists or it was not successfully renamed, an exception is thrown.

Value

Returns the pathname with the temporary suffix dropped.

Author(s)

Henrik Bengtsson

See Also

See pushTemporaryFile() for more details and an example.

[Package R.utils version 2.12.3 Index]