printf.Verbose {R.utils}	R Documentation

Formats and prints object if above threshold

Description

Formats and prints object if above threshold. The output is indented according to *enter()/*exit() calls.

Usage


## S3 method for class 'Verbose'
printf(this, fmtstr, ..., level=this$defaultLevel, timestamp=this$.timestamp)

Arguments

...
Objects to be passed to sprintf().

fmtstr
A character string specify the printf format string.

level
A numeric value to be compared to the threshold.

timestamp
A logical indicating if output should start with a timestamp, or not.

Value

Returns nothing.

Author(s)

Henrik Bengtsson

See Also

For more information see Verbose.

[Package R.utils version 2.12.3 Index]