promptAndSave.Settings {R.utils}	R Documentation

Prompt user to save modified settings

Description

Prompt user to save modified settings.

Usage


## S3 method for class 'Settings'
promptAndSave(this, saveOption="saveSettings", settingsName=NULL, ...)

Arguments

saveOption
A character string of the option used to set if user is prompted or not.

...
Arguments passed to saveAnywhere().

Details

If settings has been modified since loaded, the user is by default prompted to save the settings (if R runs interactively). To save or not save without asking or when R runs non-interactively, set option "saveSettings" to "yes" or "no", respectively. For prompting the user, use "prompt".

Value

Returns TRUE if settings were successfully written to file, otherwise FALSE is returned. An exception may also be thrown.

Author(s)

Henrik Bengtsson

See Also

*isModified(). interactive(). For more information see Settings.

[Package R.utils version 2.12.3 Index]