pushState.Verbose {R.utils}	R Documentation

Pushes the current indentation state of the Verbose object

Description

Pushes the current indentation state of the Verbose object, which is controlled by *enter() and *exit(). By pushing the state when entering a function and using on.exit() to pop the state, the correct state will set regardless of if the functions returned naturally or via an error.

Usage


## S3 method for class 'Verbose'
pushState(this, ...)

Arguments

...
Not used.

Value

Returns (invisibly) TRUE.

Author(s)

Henrik Bengtsson

See Also

For more information see Verbose.

[Package R.utils version 2.12.3 Index]