readRdHelp {R.utils}	R Documentation

Reads one or more Rd help files in a certain format

Description

Reads one or more Rd help files in a certain format.

Usage


## Default S3 method:
readRdHelp(..., format=c("text", "html", "latex", "rd"), drop=TRUE)

Arguments

...
Arguments passed to help.

format
A character string specifying the return type.

drop
If FALSE or more than one help entry is found, the result is returned as a list.

Value

Returns a list of character strings or a single character string.

Author(s)

Henrik Bengtsson

[Package R.utils version 2.12.3 Index]