Reads a single column from file in table format

Reads a single column from file in table format, which can then be used as a index-to-row (look-up) map for fast access to a subset of rows using readTable().

# S3 method for default
readTableIndex(..., indexColumn=1, colClass="character", verbose=FALSE)

Arguments

indexColumn

An single integer of the index column.

colClass

A single character specifying the class of the index column.

...

Arguments passed to readTable() used internally.

verbose

A logical or a Verbose object.

Value

Returns a vector.

Author

Henrik Bengtsson

See also

Examples

if (FALSE) {
   # File containing data table to be access many times
   filename <- "somefile.txt"

   # Create a look-up index
   index <- readTableIndex(filename)

   # Keys of interest
   keys <- c("foo", "bar", "wah")

   # Read only those keys and do it fast
   df <- readTable(filename, rows=match(keys, index))
}