Java$readUTF {R.utils}	R Documentation

Reads a Java (UTF-8) formatted string from a connection

Description

Reads a Java (UTF-8) formatted string from a connection.

Usage


## Static method (use this):
## Java$readUTF(con, as.character=TRUE, ...)

## Don't use the below:
## S3 method for class 'Java'
readUTF(static, con, as.character=TRUE, ...)

Arguments

con
Binary connection to be read from.

as.character
If TRUE, the read string converted, i.e. translated, into an R character string before returned, otherwise an integer vector representation of the Unicode string is returned.

...
Not used.

Details

Currently only 8-bit UTF-8 byte sequences are supported, i.e. plain ASCII sequences, i.e. characters that take up more than one byte are read incorrectly without any warnings.

Value

Returns a character string or an integer vector.

Author(s)

Henrik Bengtsson

See Also

readBin(). For more information see Java.

[Package R.utils version 2.12.3 Index]