reassignInPackage {R.utils}	R Documentation

Re-assigns a new value to an existing object in a loaded package

Description

Re-assigns a new value to an existing object in a loaded package.

Usage


## Default S3 method:
reassignInPackage(name, pkgName, value, keepOld=TRUE, ...)

Arguments

name
The name of the object to be replaced."

pkgName
The name of the package where the object lives."

value
The new value to be assigned.

keepOld
If TRUE, the old value is kept as attribute oldValue in the new object.

...
Not used.

Value

Returns (invisibly) the new object.

Author(s)

Henrik Bengtsson

See Also

See assignInNamespace() in getFromNamespace.

[Package R.utils version 2.12.3 Index]