removeDirectory {R.utils}	R Documentation

Removes a directory

Description

Removes a directory, and if requested, also its contents.

Usage


## Default S3 method:
removeDirectory(path, recursive=FALSE, mustExist=TRUE, ...)

Arguments

path
A character string specifying the directory to be removed.

recursive
If TRUE, subdirectories and files are also removed. If FALSE, and directory is non-empty, an exception is thrown.

mustExist
If TRUE, and the directory does not exist, an exception is thrown.

...
Not used.

Value

Returns (invisibly) TRUE, the directory was successfully removed, otherwise FALSE, unless an exception is thrown.

Symbolic links

This function can also be used to remove symbolic links to directories without removing the target. Note that neither file.remove() nor unlink() is capable of remove symbolic directory links on Windows.

Author(s)

Henrik Bengtsson

See Also

Internally unlink() is used.

[Package R.utils version 2.12.3 Index]