renameFile {R.utils}	R Documentation

Renames a file (or a directory) atomically/safely

Description

Renames a file (or a directory) atomically/safely, by also asserting that it was successfully renamed without side effects. If failing to rename and overwrite an existing file, the original file is kept.

Usage


## Default S3 method:
renameFile(pathname, newPathname, overwrite=FALSE, ..., verbose=FALSE)

Arguments

pathname
The pathname of the file to be renamed.

newPathname
The new pathname. If an existing directory and the source is a file, then the destination becomes file.path(newPathname, basename(pathname)).

overwrite
If TRUE and there exists a file with new pathname, then it is overwritten.

...
Not used.

verbose
See Verbose.

Value

Returns TRUE if the file was successfully renamed. If it failed, an exception is thrown.

Author(s)

Henrik Bengtsson

See Also

file.rename().

[Package R.utils version 2.12.3 Index]