Saves settings to file

Saves settings to file. If the settings was read from file, they are by default written back to the same file. If this was not the case, it defaults to the settings file in the home directory of the current user.

# S3 method for Settings
saveAnywhere(this, file=NULL, path="~", ...)

Arguments

file

A character string or a connection where to write too. If NULL, the file from which the settings were read is used. If this was not the case, argument path is used.

path

The default path, if no settings files are specified. This defaults to the current user's home directory.

...

Arguments passed to save() in superclass Object.

Value

Returns (invisibly) the pathname to the save settings file.

Author

Henrik Bengtsson

See also