Saves an object to a file or a connection

Saves an object to a file or a connection.

# S3 method for default
saveObject(object, file=NULL, path=NULL, format=c("auto", "xdr", "rds"), compress=TRUE,
  ..., safe=TRUE)

Arguments

object

The object to be saved.

file

A filename or connection where the object should be saved. If NULL, the filename will be the hash code of the object plus ".xdr".

path

Optional path, if file is a filename.

format

File format.

compress

If TRUE, the file is compressed to, otherwise not.

...

Other arguments accepted by save() in the base package.

safe

If TRUE and file is a file, then, in order to lower the risk for incomplete files, the object is first written to a temporary file, which is then renamed to the final name.

Value

Returns (invisibly) the pathname or the connection.

Author

Henrik Bengtsson

See also