sourceDirectory {R.utils}	R Documentation

Sources files recursively to either local or global environment

Description

Sources files recursively to either local or global environment.

Usage


## Default S3 method:
sourceDirectory(path, pattern=".*[.](r|R|s|S|q)([.](lnk|LNK))*$", recursive=TRUE,
  envir=parent.frame(), onError=c("error", "warning", "skip"), modifiedOnly=TRUE, ...,
  verbose=FALSE)

Arguments

path
A path to a directory to be sourced.

pattern
A regular expression file name pattern to identify source code files.

recursive
If TRUE, subdirectories are recursively sourced first, otherwise not.

envir
An environment in which the code should be evaluated.

onError
If an error occurs, the error may stop the job, give a warning, or silently be skipped.

modifiedOnly
If TRUE, only files that are modified since the last time they were sourced are sourced, otherwise regardless.

...
Additional arguments passed to sourceTo().

verbose
A logical or a Verbose object.

Value

Returns a vector of the full pathnames of the files sourced.

Details

Subdirectories and files in each (sub-)directory are sourced in lexicographic order.

Hooks

This method does not provide hooks, but the internally used sourceTo() does.

Author(s)

Henrik Bengtsson

See Also

sourceTo() and compare to source().

[Package R.utils version 2.12.3 Index]