stext {R.utils}	R Documentation

Writes text in the margin along the sides of a plot

Description

Writes text in the margin along the sides of a plot.

Usage


## Default S3 method:
stext(text, side=1, line=0, pos=0.5, margin=c(0.2, 0.2),
  charDim=c(strwidth("M", cex = cex), strheight("M", cex = cex)), cex=par("cex"), ...)

Arguments

text
The text to be written. See mtext for details.

side
An integer specifying which side to write the text on. See mtext for details.

line
A numeric specifying on which line to write on.

pos
A numeric, often in [0,1], specifying the position of the text relative to the left and right edges.

margin
A numeric vector length two specifying the text margin.

charDim
A numeric vector length two specifying the size of a typical symbol.

cex
A numeric specifying the character expansion factor.

...
Additional arguments passed to mtext.

Value

Returns what mtext returns.

Author(s)

Henrik Bengtsson

See Also

Internally mtext is used.

[Package R.utils version 2.12.3 Index]