subplots {R.utils}	R Documentation

Creates a grid of subplots

Description

Creates a grid of subplots in the current figure. If arguments nrow and ncol are given a nrow-by-ncol grid of subplots are created. If only argument n is given then a r-by-s grid is created where |r-s| <= 1, i.e. a square or almost a square of subplots is created. If n and nrow is given then a grid with nrow rows and at least n subplots are created. Similar if n and ncol is given. The argument byrow specifies if the order of the subplots should be rowwise (byrow=TRUE) or columnwise.

Usage


## Default S3 method:
subplots(n=1, nrow=NULL, ncol=NULL, byrow=TRUE, ...)

Arguments

n
If given, the minimum number of subplots.

nrow
If given, the number of rows the grid of subplots should contain.

ncol
If given, the number of columns the grid of subplots should contain.

byrow
If TRUE, the panels are ordered row by row in the grid, otherwise column by column.

...
Not used.

Value

Returns the matrix containing the order of plots.

Author(s)

Henrik Bengtsson

See Also

layout and layout.show().

Examples


   subplots(nrow=2, ncol=3)  # 2-by-3 grid of subplots
   subplots(n=6, nrow=2)     # 2-by-3 grid of subplots
   subplots(n=5, ncol=2)     # 3-by-2 grid of subplots
   subplots(1)               # (Reset) to a 1-by-1 grid of subplots
   subplots(2)               # 1-by-2 grid of subplots
   subplots(3)               # 2-by-2 grid of subplots
   l <- subplots(8)          # 3-by-3 grid of subplots
   layout.show(length(l))

[Package R.utils version 2.12.3 Index]