swapXY.density {R.utils}	R Documentation

Swaps x and y coordinates of a density object

Description

Swaps x and y coordinates of a density object returned by density.

Usage


## S3 method for class 'density'
swapXY(object, ...)

Arguments

...
Not used.

Value

Returns a 'density' object of the same class with elements 'x' and 'y' swapped.

Author(s)

Henrik Bengtsson

See Also

See density for estimating densities. See *draw() for plotting a density along one of the sides.

[Package R.utils version 2.12.3 Index]