useRepos {R.utils}	R Documentation

Sets package repositories

Description

Sets package repositories.

Usage


useRepos(repos=NULL, where=c("before", "after", "replace"), unique=TRUE, fallback=TRUE,
  ...)

Arguments

repos
A character vector of repositories to use. If NULL, nothing is replaced.

where
A character string specifying how to add them to the current set of repositories.

unique
If TRUE, only unique repositories are set.

fallback
If TRUE, any remaining non-specified repository value of format '...@' (e.g. '@CRAN@') than could not be recovered by other means, will be assigned to a pre-defined known value, if possible. If so, then an informative warning is given.

...
Not used.

Value

Returns a list with element 'repos' reflecting options("repos") as the options where prior to calling this function.

Author(s)

Henrik Bengtsson

See Also

withRepos().

[Package R.utils version 2.12.3 Index]