Java$writeInt {R.utils}	R Documentation

Writes a integer (32 bits) to a connection in Java format

Description

Writes one or several integer's (32 bits) to a connection in Java format so they will be readable by Java. All data types in Java are signed, i.e. a byte can hold a value in the range [-2147483648,2147483647]. Trying to write a value outside this range will automatically be truncated without a warning.

Usage


## Static method (use this):
## Java$writeInt(con, i, ...)

## Don't use the below:
## S3 method for class 'Java'
writeInt(static, con, i, ...)

Arguments

con
Binary connection to be written to.

i
Vector of integers to be written.

Author(s)

Henrik Bengtsson

See Also

For more information see Java.

[Package R.utils version 2.12.3 Index]