NullVerbose {R.utils}	R Documentation

A Verbose class ignoring everything

Description

Package: R.utils
Class NullVerbose

Object
~~|
~~+--Verbose
~~~~~~~|
~~~~~~~+--NullVerbose

Directly known subclasses:

public static class NullVerbose
extends Verbose

A Verbose class ignoring everything.

Usage

NullVerbose(...)

Arguments

...
Ignored.

Fields and Methods

Methods:

cat 	 -

enter 	 -

evaluate 	 -

exit 	 -

header 	 -

isOn 	 Checks if the output is on.

isVisible 	 Checks if a certain verbose level will be shown or not.

newline 	 -

print 	 -

printf 	 -

ruler 	 -

str 	 -

summary 	 -

writeRaw 	 All output methods.

Methods inherited from Verbose:
as.character, as.double, as.logical, capture, cat, enter, enterf, equals, evaluate, exit, getThreshold, getTimestampFormat, header, isOn, isVisible, less, more, newline, off, on, popState, print, printf, pushState, ruler, setDefaultLevel, setThreshold, setTimestampFormat, str, summary, timestamp, timestampOff, timestampOn, warnings, writeRaw

Methods inherited from Object:
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, names, objectSize, print, save

Author(s)

Henrik Bengtsson

Examples

verbose <- Verbose()
cat(verbose, "A verbose messages")

verbose <- NullVerbose()
cat(verbose, "A verbose messages")   # Ignored

[Package R.utils version 2.9.1 Index]