System {R.utils}	R Documentation

Static class to query information about the system

Description

Package: R.utils
Class System

Object
~~|
~~+--System

Directly known subclasses:

public static class System
extends Object

The System class contains several useful class fields and methods. It cannot be instantiated.

Fields and Methods

Methods:

currentTimeMillis 	 Get the current time in milliseconds.

findGhostscript 	 Searches for a Ghostview executable on the current system.

findGraphicsDevice 	 Searches for a working PNG device.

getHostname 	 Retrieves the computer name of the current host.

getMappedDrivesOnWindows 	 -

getUsername 	 Retrieves the name of the user running R.

mapDriveOnWindows 	 -

openBrowser 	 Opens an HTML document using the OS default HTML browser.

parseDebian 	 Parses a string, file or connection for Debian formatted parameters.

unmapDriveOnWindows 	 -

Methods inherited from Object:
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, names, objectSize, print, save

Author(s)

Henrik Bengtsson

[Package R.utils version 2.11.0 Index]