downloadFile.character {R.utils}	R Documentation

Downloads a file

Description

Downloads a file.

Usage


## S3 method for class 'character'
downloadFile(url, filename=basename(url), path=NULL, skip=TRUE, overwrite=!skip, ...,
  username=NULL, password=NULL, binary=TRUE, dropEmpty=TRUE, verbose=FALSE)

Arguments

url
A character string specifying the URL to be downloaded.

filename, path
(optional) character strings specifying the local filename and the path of the downloaded file.

skip
If TRUE, an already downloaded file is skipped.

overwrite
If TRUE, an already downloaded file is overwritten, otherwise an error is thrown.

...
Additional arguments passed to download.file.

username, password
character strings specifying the username and password for authenticated downloads. The alternative is to specify these via the URL.

binary
If TRUE, the file is downloaded exactly "as is", that is, byte by byte (recommended).

dropEmpty
If TRUE and the downloaded file is empty, the file is ignored and NULL is returned.

verbose
A logical, integer, or a Verbose object.

Details

Currently arguments username and password are only used for downloads via URL protocol 'https'. The 'https' protocol requires that either of 'curl' or 'wget' are available on the system.

Value

Returns the local pathname to the downloaded filename, or NULL if no file was downloaded.

Author(s)

Henrik Bengtsson

See Also

Internally download.file is used. That function may generate an empty file if the URL is not available.

Examples

## Not run: 
 pathname <- downloadFile("https://www.r-project.org/index.html", path="www.r-project.org/")
 print(pathname)

## End(Not run)

[Package R.utils version 2.12.3 Index]