file.info2 {R.utils}	R Documentation

Extract File Information (acknowledging symbolic file links also on Windows)

Description

Extract File Information (acknowledging symbolic file links also on Windows).

Usage

file.info2(...)

Arguments

...
A character vectors containing file paths. Tilde expansion is done: see path.expand().

Value

A data.frame. See file.info() for details.

Author(s)

Henrik Bengtsson

See Also

Internally, file.info() is used, which does not respect symbolic links on Windows. Instead, on Windows, Sys.readlink2() is used for such link to identify the target file and retrieve the file information on that instead.

[Package R.utils version 2.9.1 Index]