gcat {R.utils}	R Documentation

Parses, evaluates and outputs a GString

Description

Parses, evaluates and outputs a GString.

Usage

## Default S3 method:
gcat(..., file="", append=FALSE, envir=parent.frame())

Arguments

...
character strings passed to gstring().

file
A connection, or a pathname where to direct the output. If "", the output is sent to the standard output.

append
Only applied if file specifies a pathname If TRUE, then the output is appended to the file, otherwise the files content is overwritten.

envir
The environment in which the GString is evaluated.

Value

Returns (invisibly) a character string.

Author(s)

Henrik Bengtsson

See Also

gstring().

[Package R.utils version 2.9.1 Index]