Arguments$getEnvironment {R.utils}	R Documentation

Gets an existing environment

Description

Gets an existing environment.

Usage

## Static method (use this):
## Arguments$getEnvironment(envir=NULL, .name=NULL, ...)

## Don't use the below:
## S3 method for class 'Arguments'
getEnvironment(static, envir=NULL, .name=NULL, ...)

Arguments

envir
An environment, the name of a loaded package, or NULL. If NULL, the global environment is returned.

.name
A character string for name used in error messages.

...
Not used.

Value

Returns an environment.

Author(s)

Henrik Bengtsson

See Also

For more information see Arguments.

[Package R.utils version 2.9.1 Index]