System$getHostname {R.utils}	R Documentation

Retrieves the computer name of the current host

Description

Retrieves the computer name of the current host.

Usage


## Static method (use this):
## System$getHostname(...)

## Don't use the below:
## S3 method for class 'System'
getHostname(static, ...)

Details

First, this function checks the system environment variables HOST, HOSTNAME, and COMPUTERNAME. Second, it checks Sys.info()["nodename"] for host name details. Finally, it tries to query the system command uname -n.

Value

Returns a character string.

See Also

*getUsername().

[Package R.utils version 2.12.3 Index]