hsize {R.utils}	R Documentation

Convert byte sizes into human-readable byte sizes

Description

Convert byte sizes into human-readable byte sizes.

Usage

hsize(...)

Arguments

sizes
A numeric vector of sizes.

digits
Number of digits to be presented in the give unit.

units
A character string specifying type of units to use.

bytes
The string used for units of bytes without a prefix. Applied only if units="auto".

...
Not used.

Value

A character vector.

Author(s)

Henrik Bengtsson

See Also

object.size.

Examples

sizes <- c(1000^(0:8), 1024^(0:8))
df <- data.frame(size=sizes)
df$SI <- hsize(sizes, standard="SI")
df$TB <- hsize(sizes, units="TB")
df$IEC <- hsize(sizes, standard="IEC")
df$TiB <- hsize(sizes, units="TiB")
print(df)

[Package R.utils version 2.9.1 Index]