mout {R.utils}	R Documentation

Miscellaneous functions for outputting via message()

Description

Miscellaneous functions for outputting via message(). These "m*" methods work analogously to their corresponding "*" methods print(), cat(), show, str, and printf() but uses message() to output the content, which in turn outputs to standard error. The mout() method can be used for all other output methods, e.g. mout(write(x, file=stdout())).

Usage


mout(..., appendLF=FALSE)

Arguments

...
Arguments passed to the underlying output method.

appendLF
A logical specifying whether to append a newline at the end or not.

Value

Returns what the message() returns.

Author(s)

Henrik Bengtsson

Examples


  print(letters[1:8])
  mprint(letters[1:8])

  cat(c(letters[1:8], "\n"))
  mcat(c(letters[1:8], "\n"))

  str(letters[1:8])
  mstr(letters[1:8])

  printf("x=%d\n", 1:3)
  mprintf("x=%d\n", 1:3)

[Package R.utils version 2.12.3 Index]