tmpfile {R.utils}	R Documentation

Creates a temporary file with content

Description

Creates a temporary file with content that will auto delete as soon as there is no longer any references to it.

Usage

tmpfile(content=NULL, ...)

Arguments

content
A character string to be written to the file.

...
Optional arguments passed to tempfile().

Value

The absolute pathname to the temporary file.

Author(s)

Henrik Bengtsson

See Also

tempfile().

Examples

  md5 <- tools::md5sum(tmpfile("Hello world!"))
  print(md5)

[Package R.utils version 2.9.1 Index]