toCamelCase {R.utils}	R Documentation

Converts a string of words into a merged camel-cased word

Description

Converts a string of words into a merged camel-cased word, e.g. "a single espresso" is converted to "aSingleEspresso".

Usage

## Default S3 method:
toCamelCase(s, capitalize=FALSE, preserveSameCase=FALSE, split="[ \t]+", ...)

Arguments

s
A character vector.

capitalize
If TRUE, the first letter will be in upper case, otherwise it will be in lower case.

preserveSameCase
If TRUE, words that are in all upper case will remain as all same case words, e.g. acronyms.

split
A pattern used to identify words. See strsplit() for more details.

...
Not used.

Value

Returns a character vector.

Author(s)

Henrik Bengtsson

See Also

capitalize. chartr().

Examples

s <- "hello world"
print(toCamelCase(s))  # helloWorld
print(toCamelCase(s, capitalize=TRUE))  # HelloWorld
stopifnot(toCamelCase(s) == toCamelCase(toCamelCase(s)))

s <- "GEO Accession"
print(toCamelCase(s))  # gEOAccession
print(toCamelCase(s, preserveSameCase=TRUE))  # geoAccession
print(toCamelCase(s, capitalize=TRUE))  # GEOAccession
print(toCamelCase(s, capitalize=TRUE, preserveSameCase=TRUE))  # GEOAccession
stopifnot(toCamelCase(s) == toCamelCase(toCamelCase(s)))

[Package R.utils version 2.9.1 Index]