writeBinFragments {R.utils}	R Documentation

Writes binary data to disjoint sections of a connection or a file

Description

Writes binary data to disjoint sections of a connection or a file.

Usage

## Default S3 method:
writeBinFragments(con, object, idxs, size=NA, ...)

Arguments

con
A connection or the pathname of an existing file.

object
A vector of objects to be written.

idxs
A vector of (non-duplicated) indices or a Nx2 matrix of N from-to index intervals specifying the elements to be read. Positions are always relative to the start of the file/connection.

size
The size of the data type to be read. If NA, the natural size of the data type is used.

...
Additional arguments passed to writeBin().

Value

Returns nothing.

Author(s)

Henrik Bengtsson

See Also

readBinFragments().

Examples

## Not run: # See example(readBinFragments.connection)

[Package R.utils version 2.9.1 Index]